split the code
split the code that belonged to both backend & frontend.
frontend/drawing.js - scene -> backend/scene.js
frontend/controls.js - Multiway, Twoway, Fourway -> backend/controls.js
common/animation.js - SpriteAnimation -> frontend/sprite.js
frontend/extensions.js - zeroFill -> common/2D.js
frontend/extensions.js - keys, mouseButtons -> common/keycodes.js

renamed frontend/controls.js to input.js
